+++ go.mod
@@ -5,8 +5,8 @@ go 1.24.1
 replace (
 	github.com/anacrolix/torrent => github.com/erigontech/torrent v1.54.3-alpha-1
 	github.com/crate-crypto/go-kzg-4844 => github.com/erigontech/go-kzg-4844 v0.0.0-20250130131058-ce13be60bc86
-	github.com/erigontech/erigon => ./erigon
-	github.com/erigontech/erigon-lib => ./erigon/erigon-lib
+	github.com/erigontech/erigon => ./transpiler/erigon
+	github.com/erigontech/erigon-lib => ./transpiler/erigon/erigon-lib
 	github.com/holiman/bloomfilter/v2 => github.com/AskAlexSharov/bloomfilter/v2 v2.0.9
 )
