+++ golangci.yml
